requirements.txt
mastodon.py==1.8.1
gspread==5.10.0
python-dotenv==1.0.0
google-auth==2.22.0
pyjosa==1.0.3
pytest==7.4.0
pytest-mock==3.11.1